template-only-app/requirements.txt
flask
psycopg[binary]
azure-identity
azure-storage-blob
